Removed BaseSimpleRestClient as BaseRestClient will hold all the logic Updated the ReadMe to reflect the changes
